refactor: Exception 클래스들 작명, 패키지 생성 후 정리
