+++ src/main/java/racingCar/exception/count/CountRangeException.java
@@ -1,4 +1,4 @@
-package racingCar.exception;
+package racingCar.exception.count;
 
 public class CountRangeException extends IllegalArgumentException{
 	public static final String EXCEPTION_ZERO_MESSAGE = "0회 이상 입력해주세요.";
