+++ src/main/java/racingCar/exception/name/NameDuplicatedException.java
@@ -1,9 +1,9 @@
-package racingCar.exception;
+package racingCar.exception.name;
 
-public class DuplicatedNameException extends IllegalArgumentException {
+public class NameDuplicatedException extends IllegalArgumentException {
 	public static final String EXCEPTION_DUPLICATED_NAME_MESSAGE = "중복된 이름이 입력되었습니다.";
 
-	public DuplicatedNameException() {
+	public NameDuplicatedException() {
 		super(EXCEPTION_DUPLICATED_NAME_MESSAGE);
 	}
 }
