+++ src/main/java/racingCar/exception/name/NameRangeException.java
@@ -1,4 +1,4 @@
-package racingCar.exception;
+package racingCar.exception.name;
 
 public class NameRangeException extends IllegalArgumentException {
 	private static final String EXCEPTION_NAME_RANGE = "이름은 5자 이하만 가능합니다.";
