+++ src/main/java/racingCar/exception/name/NameSpaceException.java
@@ -1,4 +1,4 @@
-package racingCar.exception;
+package racingCar.exception.name;
 
 public class NameSpaceException extends IllegalArgumentException {
 	private static final String EXCEPTION_SPACE_MESSAGE = "이름이 공백으로 이뤄져 있습니다";
