+++ src/main/java/racingCar/service/RequestService.java
@@ -1,36 +1,25 @@
 package racingCar.service;
 
-import racingCar.exception.NullNameException;
+import racingCar.exception.NullInputException;
 import racingCar.utlis.Util;
 import racingCar.view.Input;
-import racingCar.view.Output;
 
 public class RequestService {
 
 	public static void requestNames() {
-		try {
-			String input = Input.inputNames();
-			validateNotNullInput(input);
-			RacingCarsService.initiateCars(input);
-		} catch (Exception exception) {
-			Output.printError(exception.getMessage());
-			requestNames();
-		}
+		String input = Input.inputNames();
+		validateNotNullInput(input);
+		RacingCarsService.initiateCars(input);
 	}
 
 	public static void requestCount() {
-		try {
-			String input = Input.inputCount();
-			validateNotNullInput(input);
-			RacingCarsService.initiateCount(Util.convertToInteger(input));
-		} catch (Exception exception) {
-			Output.printError(exception.getMessage());
-			requestCount();
-		}
+		String input = Input.inputCount();
+		validateNotNullInput(input);
+		RacingCarsService.initiateCount(Util.convertToInteger(input));
 	}
 
-	private static void validateNotNullInput(String inputString) throws Exception {
+	private static void validateNotNullInput(String inputString) {
 		if (inputString == null || inputString.isEmpty())
-			throw new NullNameException();
+			throw new NullInputException();
 	}
 }
